+++ src/Lexer/lexical_rules.py
@@ -13,5 +13,9 @@ symbols = {
     "-": "MINUS",
     "*": "STAR",
     "/": "SLASH",
+    "{": "LEFT_CURLY_BRACE",
+    "}": "RIGHT_CURLY_BRACE",
     # two char
 }
+
+keywords = {""}