+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changes
 
-## 0.1.0 [pre-alpha] (2023-02-08)
+## 0.1.0 [pre-alpha] (2023-02-12)
 
 ### Contributors
  - Jack lloyd-walters.
 
 ### Fixes
+ - Add what's new to toctree.
  - Doc links broken.
  - Docs fetch the release numbers without import.
  - Docs don't attempt to load errors.
@@ -13,10 +14,13 @@
  - Run_nox workflow correctly find version info.
 
 ### Improvements
+ - Add interpolation tests.
  - Introduce `--build-num` so tests work again.
  - Better nox tests for skiylia build number.
 
 ### New features
+ - String interpolation token + example.
+ - Lint the docs.
  - Add skiylia highlighter to documentation.
  - Introduce changelog and changelog workflow.